sensor_dashboard/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.2 